feat: Update favicon configuration with new tile color
The code changes include updating the `TileColor` property in the `browserconfig.xml` file to use a new tile color for the favicon. This commit message follows the established convention of using a prefix to indicate the type of change (feat for a new feature) and a concise description of the changes made.