feat: Add embla-carousel-react dependency for image carousel
The code changes in the package.json file add the embla-carousel-react dependency to enable the use of an image carousel on the website. This commit message follows the established convention of using a prefix to indicate the type of change (feat for a new feature) and provides a concise description of the changes made.