feat: Update global CSS to use a lighter stroke color for title
The code changes in the `globals.css` file modify the `.titleStroke` class to use a lighter stroke color for the title. This improves the visual appearance of the title by making it less prominent. This commit message follows the established convention of using a prefix to indicate the type of change (feat for a new feature) and provides a concise description of the changes made.