feat: Add @radix-ui/react-hover-card and @radix-ui/react-tooltip dependencies
This commit adds the "@radix-ui/react-hover-card" and "@radix-ui/react-tooltip" dependencies to the package.json file. These dependencies are required for the new tooltip and hover card components that have been added to the project. The commit message follows the established convention of using a prefix to indicate the type of change (feat for a new feature) and provides a concise description of the changes made.