feat: Update npm dependencies to latest versions
The code changes in the package.json file update several npm dependencies to their latest stable versions. This ensures that the project is using the most up-to-date versions of the dependencies. The commit message follows the established convention of using a prefix to indicate the type of change (feat for a new feature) and provides a concise description of the changes made.